# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstaApp API",
        "description": "API backend untuk aplikasi InstaApp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "List semua post",
                "description": "Mengambil semua postingan beserta media.",
                "operationId": "c6f2cc414b2da5257e4160d7323e44bc",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil list post"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Buat post baru",
                "description": "Membuat postingan baru oleh user yang login.",
                "operationId": "44c0e176ff8f49c4114ba82a73f017c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "caption": {
                                        "type": "string",
                                        "example": "Liburan hari ini!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil membuat post"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Detail post",
                "description": "Mengambil detail post beserta media.",
                "operationId": "f2f6c491b576bd042008143267aacb6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil ambil detail post"
                    },
                    "404": {
                        "description": "Post tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update post",
                "description": "Hanya pemilik post yang dapat mengupdate postingannya.",
                "operationId": "23d2ff3b9890023a8c9338f7820445fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "caption": {
                                        "type": "string",
                                        "example": "Update caption baru"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil update post"
                    },
                    "403": {
                        "description": "Tidak boleh update post orang lain"
                    },
                    "404": {
                        "description": "Post tidak ditemukan"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Hapus post",
                "description": "Hanya pemilik post yang boleh menghapus post.",
                "operationId": "7d5af96085e32c0bd29c22143322831d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post berhasil dihapus"
                    },
                    "403": {
                        "description": "Tidak boleh hapus post orang lain"
                    },
                    "404": {
                        "description": "Post tidak ditemukan"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post-media/{postId}": {
            "post": {
                "tags": [
                    "Post Media"
                ],
                "summary": "Upload media to a post",
                "operationId": "b32910c68f9f7a3dd3a24ae592f00ab6",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "UUID of the post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Media file (image/video)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type": {
                                        "description": "Media type (image/video)",
                                        "type": "string",
                                        "example": "image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Media uploaded"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}